requirement.txt
fastapi[standard]
uvicorn
sqlmodel
psycopg2
pyjwt
bcrypt==4.0.1
passlib[bcrypt]